Replace build with esm
