--- .babelrc
@@ -1,15 +1,0 @@
-{
-    "presets": [
-        [
-            "@babel/preset-env",
-            {
-                "targets": {
-                    "browsers": [
-                        "> 0.25%",
-                        "ie >= 10"
-                    ]
-                }
-            }
-        ]
-    ]
-}
