+++ .drone.yml
@@ -19,15 +19,6 @@ steps:
   commands:
   - npm test
 
-- name: build
-  pull: default
-  image: node:10
-  commands:
-  - npm run build
-  when:
-    event:
-    - tag
-
 - name: deploy
   pull: default
   image: node:10
