+++ package.json
@@ -2,9 +2,8 @@
   "name": "@meltwater/coerce",
   "version": "0.2.0",
   "description": "A simple javascript package for type checking an object",
-  "main": "cjs/index.js",
+  "main": "src/index.js",
   "scripts": {
-    "build": "npm run clean && babel src --out-dir cjs/",
     "clean": "rm -rf cjs",
     "docs": "documentation build src/index.js -f md > docs/API.md",
     "postversion": "git push --follow-tags",
@@ -28,11 +27,11 @@
   },
   "homepage": "https://github.com/meltwater/coerce#readme",
   "devDependencies": {
-    "@babel/cli": "7.6.2",
-    "@babel/core": "7.6.2",
-    "@babel/preset-env": "7.6.2",
     "documentation": "12.1.2",
     "eslint": "6.5.1",
     "jest": "24.9.0"
+  },
+  "dependencies": {
+    "esm": "3.2.25"
   }
 }
